XUnitFinal/XUnitConsole/Calculator.cs
using System;
using System.Collections.Generic;
using System.Text;

namespace XUnitConsole
{
    public class Calculator
    {

        public static double Addition(double num1, double num2)
        {
            return num1 + num2;
        }

        public static double Subtraction(double num1, double num2)
        {
            return num1 - num2;
        }

        public static double Multiplication(double num1, double num2)
        {
            return num1 * num2;
        }

        public static double Division(double num1, double num2)
        {
            if (num2 == 0)
            {
                return 0;
            }
            else
            {
                return num1 / num2;
            }
        }
    }
}
